Move APerformanceHintManager into its own shared library
This moves APHM from the native NDK into its own shared library for
easier test mocking and reuse by other components. It also modifies
some of those components to depend on the library instead of the NDK
directly.

The tests for the NDK were moved to the new library, since they exist to
test the logic of this code, and were ported to cover the library. The
NDK surface is still tested by a separate CTS which is not modified.

Bug: 361846430

Flag: EXEMPT PURE_REFACTOR

Test: atest HintManagerServiceTest
Test: atest PerformanceHintNativeTestCases
Test: atest PerformanceHintManagerTest
Change-Id: Iac57f0fc2c8c1cc081fb04d2a932196b9f365189