Crash when Binder would have to drop a BR_REPLY
If a BR_REPLY comes in when we're not expecting a reply, this is a bug
in binder where the driver and library don't agree on when a reply
should be expected. This should not silently drop the reply, because
otherwise all you see is something blocked forever waiting for a reply
that was swallowed up here.

The linked bug is a bug in Starnix binder that would have been much more
obvious if this log was here.

Bug: 417917758
Test: TODO
Flag: EXEMPT DEBUG
Change-Id: I104cc480bbaeaf1c3c974566805fc8b0b1d759dd