Shift key remapping to KL based remapping
Previous remapping was implemented in KCM and was based on keycode
which works well for keyboards since KL files are usually never
changed and uses Generic.kl. But for controllers, remotes, etc.
KL files are changed which makes it difficult to define remapping
based on keycodes.

Use "location" on L.H.S. for remapping making things more predictable
when device KL files can also be changed for the peripheral.

"Location keycode" refers to the standard location a key is present
in a default QWERTY device (i.e. can be used to reverse map to scan code
using Generic.kl file). Use the scancode to define an overlay remapping
directly in KeyLayout file which is the standard place for defining scan
code to keycode mappings.
Accordingly also change the "getKeycodeForKeyLocation" method call.

We plan to use this "location" to "keycode" mappings for game controller
remapping too and this refactor required to reuse the existing remapping
infra

Bug: 435503326
Flag: EXEMPT BUGFIX
Test: atest ModifierKeyRemapping
Change-Id: Iba3ad2375eaf28b842ab51778741319e0ec7eb28