Improve RRect intersection and crop insetting
Corrects the RRect intersection logic to check if both sides of the Rect
satisfy the requirement for minimum width and minimum height. For the
insetCrop, we should use  maximum corner radius dimensions for insetting.
If the bounds do not intersect this maximally-inset crop, it provides a
definitive guarantee that it's also not touching the original RRect's
rounded boundaries.

Test: atest librenderengine_test SurfaceFlinger_test
Bug: 444721231
Flag: EXEMPT BUGFIX
Change-Id: Id63a299d468461eaa9188661fb036b9a6f131d6d