Add binder netlink report support
This adds a class BinderNetlink to talk to the kernel binder netlink
driver. It allows AMS to listen to the kernel driver for certain
kinds of abnormal binder transactions and take corresponding actions.

The class is placed in its own library, so it has no impact on binder
clients.  Only system_server uses it.

The following binder tests are added to test the kernel driver:

* BinderNetlinkOpen: open and close the binder generic netlink socket
* BinderNetlinkMonitor: read the binder generic netlink socket
* BinderNetlinkXxxReply: test various kinds of transactions

The tests report "Ignored" if the kernel does not have binder netlink
support.  At this time, GKI kernels at or above 6.12.38 have binder
netlink support.

Flag: android.os.binder_netlink_enabled
Bug: 332610283
Test: boot Pixel with mainline kernel and check logcat
Test: atest
 * binderLibTest

Change-Id: Ie7dbc47b62d01a87da44de74b86e5424c31ea9ce