Determine display power mode from all frame targets
Work duration is not reported to the power hint session when
the front internal display turns off, but other frame targets
are still on. This poses an issue in scenarios where a connected
display in desktop mode is rendering on the external display, but
the device screen turns off. The conesequences of not reporting
work duration in these instances includes the CPU and GPU
frequencies not appropriately adjusting to the workload and leading
to jank.

Check if any SF frame targets are powered on and keep session
reporting on if so.

Bug: b/434572841
Test: connect to an external display and let the device screen time out.
Verify in perfetto that work duration is reported and CPU/GPU frequencies are changing
Flag: EXEMPT BUGFIX

Change-Id: Iafcdf642a5d08bbb53c119c4c0227f6ebd541831