Reset mFrozen->initialStateReceived in onLastStrongRef
This fixes a minor issue where a newly registered listener might receive
a stale state as the initial state. This isn't too big of a problem
because the actual initial state (if different) would soon be delivered
as a follow-up state change. Still it would be more correct to reset
this variable.

I don't think this issue causes the ref count problem that we are
investigating now.

Bug: 425638507
Test: TH
Flag: build.RELEASE_LIBBINDER_DEFER_BC_REQUEST_FREEZE_NOTIFICATION

Change-Id: I44b7dc0c7c18dc5aa556aa0e570518928d1324da