SF: Avoid pacesetter churn on foldables
Currently, if all displays are powered off, the Scheduler selects the
first display that was added as the pacesetter display. This, for
foldables, causes churn in pacesetter when folded, as the outer
display, which was last on and thus the pacesetter, would lose its
designation when powered off since the inner display is the first
display to be added. To avoid this churn, and another change when the
outer display powers on, This CL changes the pacesetter selection
logic to prefer the current pacesetter if all displays are powered off.
The inner display will maintain its pacesetter display status on boot
as it is the first display added.

Flag: EXEMPT BUGFIX
Bug: 447180790
Test: New SchedulerTest
Test: Manual via dumpsys w/ on comet on boot
Test: Manual via dumpsys w/ on comet folded then powered off (AOD OFF)
Change-Id: Ie72632b09c9d7fde6f4a1ee97c11d588b60ec4da