SF: Apply new vsync schedule and phase config on pacesetter promo
Scheduler::register/unregisterDisplay() triggers pacesetter selection
as non-internal displays can become a pacesetter now.

Whenever a pacesetter changes, Scheduler::updatePhaseConfiguration()
should be called to update the vsync configuration and callbacks
(among others) with the new vsync. However, updatePhaseConfiguration()
was only called in Scheduler::designatePacesetterDisplay(), which is
only called during display and power mode changes. This meant that if a
pacesetter display was unregistered, the pacesetter designation would
change, but updatePhaseConfiguration() would not be called, which led
to all sorts of scheduling issues.

This CL adds updatePhaseConfiguration() to promotePacesetterDisplay(),
as well as applyNewVsyncSchedule() and VsyncConfiguration::reset().

Flag: EXEMPT BUGFIX
Bug: 430098695
Test: Manually test - unplugging the pacesetter display w/ chrome
Change-Id: I514d446d5bfd4a7bb7590455734318485cdcece1