InputVerifier: check down times (attempt 2)
Specifically, that the down time of an event never goes backwards
(decreases).

This should catch issues like the one with touchpad button press events
that was recently fixed [0].

I considered also checking that the down time only ever changes on a
DOWN event, but I discovered that this causes crashes for mouse clicks,
where CursorInputMapper sets the down time on the HOVER_MOVE event
preceding the DOWN for the click. In discussion, some on the team
thought this was desirable behaviour so we're keeping it.

In the first attempt at this CL, I checked that the down time of DOWN
events was equal to their event time, but this caused trouble with:

* the accessibility input filter for magnification (b/429719275), which
  re-injects DOWN events with increased down times to avoid incorrectly
  triggering long presses; and
* some tests that use UiAutomator, which sometimes injects DOWN events
  with down times earlier than their event times. (b/450221584)

This second issue at least is something we should fix, but may have side
effects on other tests, so for now I'm dropping that check altogether.
In future I'll try to get UiAutomator fixed so that we can at least
assert that a DOWN event should have a down time greater than or equal
to its event time.

[0]: change Ia56d48fda658c8344618734750c95e9d3d24f897
[1]: $ adb shell 'stop && setprop log.tag.InputDispatcherVerifyEvents \
                          DEBUG && start'

Bug: 447603159
Test: with the flag enabled, enable verification [1], and use various
      input devices (touchpad, touchscreen, mouse, stylus) to check that
      the verifier doesn't crash
Test: remove the fix for b/447603159 [0], enable verification [1], and
      check that clicking on a touchpad causes a crash only when the
      flag is enabled
Test: $ atest --host libinput_rust_test
Test: $ atest --host inputflinger_tests
Test: $ atest 'CtsAccessibilityServiceTestCases:android.accessibilityservice.cts.FullScreenMagnificationGestureHandlerTest#testZoomOnOff'
Flag: com.android.input.flags.touchpad_down_time_fix
Change-Id: I7ee66b5785715a2b81af8985ca0bcd74b14bd46d