InputVerifier: check down times
Specifically, that:

* the down time of an event never goes backwards (decreases); and
* when we see a DOWN event, its down time matches its event time.

This should catch issues like the one with touchpad button press events
that was recently fixed [0].

I considered also checking that the down time only ever changes on a
DOWN event, but I discovered that this causes crashes for mouse clicks,
where CursorInputMapper sets the down time on the HOVER_MOVE event
preceding the DOWN for the click. In discussion, some on the team
thought this was desirable behaviour so we're keeping it.

[0]: change Ia56d48fda658c8344618734750c95e9d3d24f897
[1]: $ adb shell 'stop && setprop log.tag.InputDispatcherVerifyEvents \
                          DEBUG && start'

Bug: 447603159
Test: with the flag enabled, enable verification [1], and use various
      input devices (touchpad, touchscreen, mouse, stylus) to check that
      the verifier doesn't crash
Test: remove the fix for b/447603159 [0], enable verification [1], and
      check that clicking on a touchpad causes a crash only when the
      flag is enabled
Test: $ atest --host libinput_rust_test
Test: $ atest --host inputflinger_tests
Flag: com.android.input.flags.touchpad_down_time_fix
Change-Id: Ie91316e6a8c19b877740f90d93dc2160f1caeaec