Allow cross-display mouse streams in the verifier
As part of cross-display cursor support, the pointer choreographer now
works in topology-aware mode. That means that the mouse cursor is able
to move from one display to another without starting / stopping the
stream.

For example, HOVER_MOVE (display 1) -> HOVER_MOVE (display 2) from the
same device id is an allowed transition. Likewise, MOVE (display 1) ->
MOVE (display 2) (while mouse button is pressed) is also allowed. The
HOVER case was likely already working earlier, because we implicitly
allow HOVER_MOVE to start a new mouse hovering gesture. However, the
ACTION_MOVE case was not.

In this CL, we allow the MOVE case to be verified.

This is accomplished by changing the way we select the relevant
verifier. For the mouse/touchpad events - the primary display id
verifier will be used instead of the "target" display verifier.
For all other types of events, the associated display verifier will
continue to be used.

This works around the issue of having "inconsistent" cross-display
streams.

When this change was implemented, I noticed that the tests started
failing for another reason - the events were simulated incorrectly for
some tests. They used to be injected in the beginning of the tests
(injection verifiers follow a different path), and notifyMotion was used
after. This means that the stream was already inconsistent coming into
the dispatcher. This is also fixed in this CL by converting all of the
involved tests to use 'notifyMotion'. In general, we should be using
notifyMotion and get away from the injection path.

Bug: 271455682
Test: TEST=inputflinger_tests; m $TEST && $ANDROID_HOST_OUT/nativetest64/$TEST/$TEST --gtest_repeat=1 --gtest_filter="*"
Flag: EXEMPT TEST_ONLY
Change-Id: I0aacd1f98a84bf614b870115095adb95e48af1a1