Fix fakeOutputDataspace transfer mask comparison
`needsCustomOETF` should be true when fakeOutputDataspace &
HAL_DATASPACE_TRANSFER_MASK is non-zero AND also equal to
HAL_DATASPACE_TRANSFER_GAMMA2_2.

As other custom OETFs get added, this could change to be `!= 0`, but for
now given the restriction to a 2.2 gamma, this fixes the check to be ==.

Without this fix, the `gamutTransform` was being added to the
`in_colorTransform`'s uniform value when it shouldn't have been.

Flag: EXEMPT BUGFIX
Bug: b/448309407
Change-Id: Ib874bbfb4437cf2947d0a8b7c7891d3346721a38