TouchpadInputMapper: disable 3-finger swipe detection in RELATIVE mode
When in RELATIVE capture mode, we don't do anything with three- or
four-finger swipes, so we may as well tell the Gestures library not to
bother sending them.

In the design doc, I'd planned to set the "Pinch Enable" gesture
property to false for similar reasons, but this turned out to be a bad
idea, as pinches then got detected as cursor movements following one of
the fingers.

Test: check the property value changes after entering and leaving
      RELATIVE capture mode, using dumpsys:
      $ adb shell dumpsys input | grep "Three Finger Swipe Enable"
Test: after capturing in RELATIVE mode and releasing capture, check
      three- and four-finger swipes still work
Test: $ atest --host inputflinger_tests:TouchpadInputMapperTest
Bug: 403531245
Flag: com.android.hardware.input.pointer_capture_modes
Change-Id: I322b68d011c81a680fc83bc60038825e102ece4d