SurfaceFlinger: Add mirrorLayerStack and plan for true mirrorDisplay
The original implementation of mirrorDisplay does not mirror the display
but the display's layer stack which can change, and that does not
produce the desired behavior in screenrecord.cpp because if the
display ID is explicitly set by the user, it does not make sense to
switch the recording to a different display.

To clarify this behavior and allow for a future, correct implementation,
this commit renames the original function to
mirrorLayerStack. This improves API clarity, clearing the way for
a future, correct mirrorDisplay method that genuinely mirrors by
display ID.

Bug: 281084746
Flag: EXEMPT BUGFIX
Test: atest SurfaceFlinger_test
Change-Id: I1e9752e220ee3fd86643d3c9040afbb4c57a06bc