Add mirrorDisplay API to mirror by stable display ID
The existing mirrorLayerStack API creates a mirror of a display's
current LayerStack. If that LayerStack is later moved to a different
physical display (e.g., on a device fold/unfold or external display
change), the mirror will follow the LayerStack, not the original
physical display.

This patch introduces a new API,
SurfaceComposerClient::mirrorDisplay(DisplayId), which allows a
client to create a mirror that is bound to a stable, physical
DisplayId.

Bug: 281084746
Flag: EXEMPT BUGFIX
Test: atest libsurfaceflinger_unittest
Change-Id: I02baf1cb75a69927124afdc435833955ddd6b5e6