usbauthservice: Allow Fibocom FM101 modem
The Fibocom FM101 modem is a USB device that provides network connectivity.
To ensure it is functional during initial device setup and on the login screen
(before a user is fully logged in), it needs to be explicitly allowed by the
USB authentication policy due to b/453040973.

This change adds a rule to `desktop_auth_policy.conf` to allow the modem
with VID:2cb7 and PID:01a2 when the system is in the `Setup` or `Booted` state.

Bug: 453040973
Flag: EXEMPT BUGFIX
Test: Manually verified that the Fibocom FM101 modem is functional on the setup and login screens.
Change-Id: Id6d3a47c5c8084848c6d7505db0e8d9bb737e697