TouchpadInputMapper: Disable acceleration in RELATIVE mode
Acceleration curves are disabled for captured mice, so do the same for
touchpads.

The MotionArgs created by RelativeModeGestureConverter will now have
their motion values in mm, but a later stage in the pipeline is scaling
these down (I suspect to account for screen density). This will be
addressed in a later CL.

[0]: $ adb shell dumpsys input | grep -E "(Pointer|Scroll) Accel Curve"

Test: using dumpsys [0], check the curves change appropriately when
      entering and leaving capture mode
Test: change the pointer speed setting while in RELATIVE mode:
      $ adb shell settings put system touchpad_pointer_speed -7
      ...then check that the curves haven't changed in dumpsys [0], but
      that the new speed does take effect once pointer capture ends
Test: $ atest --host inputflinger_tests:TouchpadInputMapperTest
Bug: 403531245
Flag: com.android.hardware.input.pointer_capture_modes
Change-Id: I04f1506f140defa70321e6050f0fb51705f73091