Add client listener registration to SensorService
This change introduces a mechanism for clients to register a listener
with the SensorService.

A new AIDL interface, ISensorClientListener, is added. The SensorManager
and SensorService are updated with a new registerClientListener method.

This allows the SensorService to keep track of its clients. This is a
foundational step towards monitoring client status, such as process
death or frozen states.

Test: dumpsys sensorservice with cmd activity freeze and see the
expected result
Flag: android.hardware.flags.suspend_sensor_event_delivery_on_frozen_pid
Bug: 409841081

Change-Id: Ia236e9ccb34e1f3303a4fec5fd49b6a79c6fe8df