usbauthservice: Allow USB HOST TO HOST ADB Cable
Allow USB ADB HOST TO HOST cable. To ensure it is functional during
initial device setup and on the login screen (before a user is fully logged in),
it needs to be explicitly allowed by the USB authentication policy.

This change adds a rule to `desktop_auth_policy.conf` to allow the USB ADB
HOST TO HOST cable with VID:0x18d1 and PID:0x205a when the system is in the
`Setup` or `Booted` state.

Bug: 456236709
Flag: EXEMPT BUGFIX
Test: Manually verified that the ADB HOST TO HOST cable is functional on the setup and login screens.
Change-Id: I8f892c7d09ec583805f22c10f845f00df146350f