Exclude rotated panels from DPU readback
...unless the client is okay with the display installation orientation

DPU readback always occurs in the panel's installed orientation. But the
screenshot returned to clients needs to be returned in the "logical"
orientation. This mismatch can cause animations involving screenshots to
flicker.

Resolve this by forcing the GPU path for screenshots. This costs more
power, but at least it won't look "wrong".

Bug: 455501983
Flag: EXEMPT BUGFIX
Test: Hack up a device with a 180 degree install orientation and observe
the power off animation

Change-Id: I8c2bd24b6148eb0ad51f92457364177771f744ad