Support creating offscreen layer explicitly
Currently, if the caller is system uid or has ACCESS_SURFACE_FLINGER
permission, a new surface with null parent will be added to root.
So add a flag to disable the default behavior. Then the caller can
control when to move the surface to target layer.

Bug: 308662081
Flag: com.android.window.flags.use_client_surface
Test: CtsWindowManagerDeviceWindow
Change-Id: I0a0a2e2c290b9243cbc61af0817d141122c91b20