Fix handling of unsuccessful finish events on cancel
When cancelling a move stream, consisting of move events, there
is an edge case where moves can be left unfinished, if the
finish event returns a non-OK status including when the socket
is full, or when the device is overloaded. This CL aims to fix it
by properly catching these non-OK statuses and returning these
events to the InputReceiver caller such that they can be added to
the outboundQueue to be retried.

Bug: 422114163
Test: atest BatchedInputEventReceiverTest#testCancelledMoveEventsFinishWhenSocketFull
Flag: com.android.input.flags.fix_input_anr_by_send_message_exception
Change-Id: Ia2369528dbae003a0dbef432774495e700948273