Allow non-trusty use of vendor stability
This is needed when libbinder_ndk is built as a dependency of
libbinder_rs for use outside of Android to communicate over stable
interfaces with Android. This is the case for TEE implementations of
trusted hals. This replaces the __TRUSTY__ preprocessor macro with a
more general __BINDER_VENDOR_STABILITY__ macro. Within the Android tree
this is only set by trusty, but it's expected that other TEE OSes will
use it too.

Bug: 456773147
Flag: EXEMPT PURE_REFACTOR
Test: cd frameworks/native/libs/binder && atest
Change-Id: Ia03bdf473f0898e0cf45a2fa64acba5be0dd817f