InputDispatcher_test: use HW timeout multiplier when testing key repeat
We were seeing flakes in FocusedWindow_StopsKeyRepeatAfterUp where
multiple key repeat events were being produced, causing
sendAndConsumeKeyUp to fail because it consumed a repeated DOWN event
rather than the UP. The timestamps of the events show the second
repeated down occurring 82ms after the initial DOWN. Since the flakes
were on virtual devices that tend to be quite slow, we think performance
issues are delaying the UP event long enough for two repeats to occur.

Ideally, we would change these tests to have a test-controlled clock, so
that they are not dependent on the physical time at all, but for now
let's use the hardware timeout multiplier to extend the timings here
when on slow hardware.

This is a speculative fix, as I have not been able to reproduce the
flakes locally.

Bug: 456335107
Test: $ atest --rerun-until-failure=100 \
            inputflinger_tests:InputDispatcherKeyRepeatTest
Flag: EXEMPT TEST_ONLY
Change-Id: I6432789e88b97400431df493f8263de9315997ca