SF: Disable producer throttling for client composition (no flag)
This CL removes the flag com.android.graphics.surfaceflinger.flags.sf_disable_producer_throttling_for_client_composition and makes the fix merged in
commit 5d99be114e0b529fa7ff8d4fcf2d079ae090aad2 to be always enabled.

The fix is addressing a performance regression that was introduced in commit cd6958694c8c1fff71c98d43616d3f43ad8be158 which is not behind a flag, so the fix doesn't need a flag either. The flag was used for soaking time in staging which was completed.

Bug: 454707399
Flag: EXEMPT FLAG_REMOVAL
Test: trunk_staging soaking
Change-Id: Ida20b4bc834d6c57278c2c75879dc8a7037aea5b