Rename rounded corner radii and fix client composition logic
The variable names in RoundedCornerState were ambiguous. For example,
`radii` did not clearly state whether it referred to radii drawn by
SurfaceFlinger, requested by the client, or drawn by the client.

This led to a bug where we were checking effectiveRadii in OutputLayer and forcing client composition when SF is not drawing rounded corners. The hasRoundedCorners() function in LayerFE is updated to check the sfDrawnRadii instead.

Bug: 456602366
Flag: EXEMPT BUGFIX
Test: atest SurfaceFlinger_test libsurfaceflinger_unittest
librenderengine_test; manual WM smoke test

Change-Id: Ie10fe32fa0fc755cdabef3f53df88280d1edf9bc