Simplify the destructor
The oringinal design is not needed since the wp to sp promotion
should protect the binder thread to access mLooper if mService cannot be
promoted. The key here is that mLooper might be nullptr if the device
doesn't have any sensors on it.

Test: abtd on platform that does not have sensors
Bug: 409841081
Flag: android.hardware.flags.suspend_sensor_event_delivery_on_frozen_pid
Change-Id: I8a25757be1b2d544e01e736deffd3d64aacb4494