Fix a gles driver unloading bug
The function should_unload_system_driver() is designed to allow
apps to unload the system driver preloaded by zygote: if zygote
preloads the GLES driver, app process forked from zygote will have
system GLES driver loaded, too, and if the app requests to use a
different GLES driver, should_unload_system_driver() will return true
and the system driver will be unloaded. The unloading of system driver
is only allowed once per app process. After that the process will
proceed to load the alternative driver: ANGLE or updatable driver from
apk. And the rest of the app process is expected to use the aternative
driver.

This process is broken if the zygote does not preload the GLES driver,
the app process requests to load alternative driver, and the app has two
threads trying to call "Loader::open()". When the second thread calls
"Loader::open()", it sees that there is already a driver loaded
(loaded by the first thread), and proceeds with
should_unload_system_driver() check. The function sees that the second
thread needs an alternative driver, and the unloading hasn't happened
for the process yet, so it returns true and unload the driver that is
still used by the first thread, causing first thread to crash.

To fix this, we should make sure that when should_unload_system_driver()
is called, the driver that is already loaded is not the alternative
driver. If it is the alternative driver, it means this process has
already loaded the driver it needed and the current thread should
proceed with the same driver.

Bug: b/448201398
Bug: b/399309085
Test: make, flash
Flag: com.android.graphics.egl.flags.do_not_unload_driver_from_updatable_apk
Change-Id: Ieb8d79a9636787ad64c0227c5ca0a839fe6f411f