Changes dumpstate to use `pidfd_open` and `poll` to monitor subprocesses
The current code uses `sigtimedwait` and `waitpid` to achieve the same
functionality. However, an approach based on signals needs very
specific division of labor in a multithreaded program to work correctly.
While this didn't seem to be an issue in dumpstate in practice, it's
still not guaranteed to work.

We stepped onto the problem in a context that's explained in detail in
the bug noted below. And we thought to contribute an improvement on the
state of the art.

The new code uses `pidfd_open` and `poll` instead of signaling, to
ensure that any process state observed must come from the process
whose PID is used in `pidfd_open`. This removes the question of
which threads observe which signals from subprocesses, since state
changes are always scoped to specific PIDs.

The change also deliberately minimally modifies the code paths from
their original form, and all log messages are unchanged. The tests
in `dumpstate_test` confirm that the semantics of DumpstateUtil are
unchanged. I added a regression test for the issue we encountered,
and confirmed that it fails for us before the change, and passes
after the change.

Flag: NONE no functional changes
Tested: `atest dumpstate_test` and manual check on a device
Bug: b/450286609
Change-Id: I54907036d22c11713550e265577d6a856ba3265a