Add no_std compatible ParcelFileDescriptor
Binder and RpcBinder allow for file descriptors to be passed within
Parcelable objects. While the Unix-like notion of file descriptors may
not exist in environments that need a no_std library, several operating
systems do implement an analogous concept of a "handle" that can be
given to user space to represent some data that is managed by a kernel.
This exists in Trusty as well as other TEE operating systems that are
the first target users of libbinder_rs_nostd.

This change introduces a no_std compatible ParcelFileDescriptor that is implemented as a c_int. It also exercises the serde capabilities of the no_std lib with no_std versions of the existing serialization tests.

Bug: 425915863
Test: atest aidl_integration_test
Test: atest rustBinderNoStdSerializationTest
Test: atest rustBinderSerializationTest
Flag: NONE refactor for std and nostd guarded by cfg

Change-Id: I09b8206d2f316af2c217fcfcc24c4280d0cbc77
Change-Id: I09b8206d2f316af2c217fcfcc24c4280d0cbc776