Dump dispatcher state when bad events are received
Instead of just crashing with an error message, add a dump of the
dispatcher so that we can better learn how the device got into this
state.

Generally, the notifyMotion call should not be dependent on the
dispatcher state, and it would normally be an issue in a lower level,
like InputReader or PointerChoreographer (or the injection path). Still,
the dump will help us understand what events preceded this.

Bug: 459776375
Test: none
Flag: EXEMPT DEBUG
Change-Id: Ib81816562d4adee1ea76c3aa4d662b02841b7df8