Add tests for socket capacity for Motion and Finished Messages
It is useful to know the capacity of the InputChannel for various events
as it can be used to both understand and test scenarios where
WOULD_BLOCK leading to edge cases. Concretely this occurs in the
related bug where a WOULD_BLOCK status of a finished message was not
handled, leading to ANRs under high load.

Bug: 36766914
Flag: EXEMPT TEST_ONLY
Test: atest InputChannelTest
Change-Id: If4ba28b95e115fcbc89d6ab9cc0b914d9143114e