Undo effect of protected_if_client flag.
This flag made it so that we only switched into protected rendering if
there was some effect that was forcing a protected layer into client
composition.

But, the HWC can arbitrarily force *any* layer into client composition,
including protected layers. For example, due to resource exhaustion.
That means this check will sometimes allow for rendering protected
layers onto an unprotected buffer. In the worst case this crashes, but
in the best case we'd have to blackout the input layer which is visually
jarring.

Moreover, the motivation for this flag doesn't seem to be correct. It
might be somewhat less efficient to sometimes allocate a protected
framebuffer if the actual scene sent to the GPU doesn't contain
protected layers, but devices need to handle the path where a protected
layer _is_ sent to the GPU anyways, so we needn't be so conservative. If
this is truly a problem, then we could move buffer allocation into
renderengine directly so that renderengine can analyze the input layers
it actually receives and decide on an allocation policy. But that's a
larger change :)

Bug: 451657059
Flag: EXEMPT BUGFIX
Test: Switch from protected video into camera

Change-Id: Ia59e05c8b7e33a7cb5822b70a7fbb882472e0038