[Desktop] Add desktop-specific bugfix flag to enable color correction
Desktop rollout of color correction change will be at a separate pace to
the rest of the platform. Add a desktop-specific flag to control the
same feature as enable_color_correction_bugfix. I could not find a way
to re-use the existing flag for a desktop-first rollout while leaving it
in trunkfood on platform.

Bug: 460493928
Test: atest DaltonizerTest
Flag: com.android.graphics.surfaceflinger.flags.enable_color_correction_desktop_bugfix
Change-Id: I822ad53c6775b3fcf78f28d63767abcd863070fc