BufferQueueProducer: Set additionalOptionsGenerationId in attachBuffer
When a buffer is attached to the buffer queue, initialize the slot's
additionalOptionsGenerationId with the BufferQueue core's value.

This is a workaround for cases where the GraphicBuffer is allocated
outside of the BufferQueue. With this workaround, BufferQueue won't
reallocate buffers when the client is doing attachBuffer->dequeueBuffer.

Test: atest SurfaceViewPreviewTest#testSurfaceSet, BufferQueue_test
Flag: EXEMPT bugfix
Bug: 461441941
Bug: 461073557
Change-Id: I474824bcfef59c4199fc898c8154ca3994767a90