API support to toggle client-drawn rounded corners optimization
We want to apply client-drawn rounded corners optimization only
when the corner radius has reached a steady-state. Otherwise, this
can lead to unexpected clipping of screenshots taken at the beginning
of a transition or when resuming an app after being dismissed.

This change introduces the `disableClientDrawnRadii` flag to LayerSnapshot.RoundedCornerState.
The flag is set on the Task layer and is inherited by all children. When the flag is true, SurfaceFlinger will report 0.f radius to the client to disable client rounding.

Bug: 457677956
Flag: com.android.graphics.surfaceflinger.flags.set_client_drawn_corner_radii
Test: atest SurfaceFlinger_test libsurfaceflinger_unittest libgui_test

Change-Id: I92ae23e0abea5289da029883fa4657da88287c62