Fix how display color matrices are applied in renderengine
Clients that do display-wide transforms send a 4x4 color matrix to manipulate the post-composited rgb colors. But, Skia's color matrix is *5x4* to allow for manipulating alpha.

Previously, we translated from a 4x4 matrix into a 5x4 SkColorMatrix by appending a zeroed-out column to the matrix, then applied the matrix to every canvas draw. But that doesn't work well with alpha-compositing; an alpha that's less than 1 will decrease the translation component of the 4x4 transform to the point of creating colors that clamp. More specifically, the alpha of the source color should not affect the transformed result, because the 4x4 color matrix is defined by HWC to be applied *post-blend*, where the alpha channel of the frame buffer is always 1.0.

Fix this by instead setting the fourth column of the SkColorMatrix to (0, 0, 0, 1) to represent that the color transform is unaffected by the source color's alpha, and set the fifth column of the SkColorMatrix to the 4th column of the display's transform (with the last element zeroed out), so that the translation component for color inversion still works as intended.

Bug: 432040449
Flag: com.android.graphics.surfaceflinger.flags.color_transform_translation
Test: Device that forces display color transforms onto the GPU apply a11y effects correctly.

Change-Id: I25fd9d2ab0704b7085b420f7660004f0c473e69e