TouchpadInputMapper: add regression test for reset consistency
While writing this test, I gained a better understanding of what was
happening in the "real world" to cause an inconsistent HOVER_EXIT event
to be reported after the reset (as opposed to the less-realistic gesture
sequences used in the UncapturedGestureConverter unit tests). I believe
that the following sequence of actions is causing the issue in the wild:

1. Scroll on the touchpad with two fingers, then lift the fingers. This
   causes UncapturedGestureConverter to think that there could be a
   fling in progress.
2. Reset the touchpad
3. Do anything that triggers a fling stop (e.g. single-finger movement,
   two-finger swipe, ...)

The fling stop gesture after the reset was being converted into a
HOVER_EXIT–DOWN–CANCEL–HOVER_ENTER sequence by
UncapturedGestureConverter (which at that point thought a fling could be
in progress). That HOVER_EXIT then caused the crash.

Test: $ atest --host inputflinger_tests:TouchpadInputMapperTest
Test: revert the fix (change Ib6f3d8549e4444e6d4da2cbcb519693e65503e50),
      run the test, and check it fails with a verifier error about
      HOVER_EXIT
Bug: 458469793
Flag: EXEMPT TEST_ONLY
Change-Id: I3c200e0e5b32f7e53da057bd2cb8d09a31d685ec