TouchpadInputMapper: fix inconsistency when capturing with button down
Because we were only resetting UncapturedGestureConverter when we left
captured mode, if you entered capture mode with a button down, an UP
would be sent when capture was released again, which was both
unrepresentative of what was actually happening and causing a verifier
crash. To fix this, reset UncapturedGestureConverter when we enter
capture mode.

Test: $ atest --host inputflinger_tests:TouchpadInputMapperTest
      (and also run it without the fix, to check the new test fails)
Test: with inbound event verification enabled, enter a world in
      Minecraft. Open the menu by pressing Esc, then click "Resume"
      using the touchpad. Open the menu again, move a finger on the
      touchpad and check that no crash occurs.
Bug: 460769015
Flag: com.android.hardware.input.pointer_capture_modes
Change-Id: If2b403cedb013e900504e9b996e3ca696909b172