Move the call to updateAngleFeatureOverrides() to Loader::open()
Currently, updateAngleFeatureOverrides() is called in
GraphicsEnv::setAngleInfo(), which is only executed for Java apps.
However, the ANGLE feature overrides need to be configured for both
system services and Java apps so ANGLE behaves the same for every type
of process, including which EGL/GLES extensions are exposed by ANGLE.

The difficulty is that updateAngleFeatureOverrides() issues Binder
calls, which Zygote cannot do. To work around this, take advantage of
the fact that Zygote preloads the GLES driver so that when cnx->dso is
valid in Loader::open(), then Zygote must have completed initialization
and the executing process must be a fork of Zygote (and can issue Binder
calls). Additionally, check cnx->angleLoaded to restrict calling
updateAngleFeatureOverrides() to only when ANGLE is the GLES driver for
the running process.

Bug: b/461292354
Test: atest ActivityManagerGetConfigTests
Test: atest IntelAngleEnd2EndTestCases:ImageTest
Test: cmd activity get-config --device | grep YUV
Flag: com.android.graphics.egl.flags.update_angle_feature_overrides_in_loader_open

Change-Id: I9d133e65c7a4c5fe1f3cee3404643bea11db6ff5